feat(testing): implement comprehensive testing infrastructure
## 🎯 Major Achievements
- **44/44 tests passing** ✅
- **SlotMachineWidget**: 11/11 comprehensive widget tests
- **Unit Tests**: 33/33 model tests (Ingredient, UserSettings)
- **Test Infrastructure**: Complete setup with working patterns

## 🧪 Widget Testing Implementation
- Widget structure and layout verification
- Responsive design testing (mobile/tablet/desktop)
- Content validation and accessibility testing
- State management without animation timer issues
- Proper TestApp wrapper with Riverpod integration

## 🔧 Technical Improvements
- Fixed test compilation issues and type conversions
- Enhanced test data consistency across all models
- Resolved JSON serialization in test helpers
- Updated .gitignore to prevent large file commits
- Fixed int to double conversion in RecipeCard widget

## 📋 Next Agent Handoff
- Created TESTING_HANDOFF_NEXT_AGENT.md with complete context
- Identified immediate priority: Fix Hive/Provider mocking
- Updated task list with small, focused tasks
- Provided research requirements (Context7, Gemini MCP)

## 🚧 Current Status
- Testing infrastructure: ✅ Complete and working
- SlotMachineWidget tests: ✅ 11/11 passing
- Unit model tests: ✅ 33/33 passing
- SpinModeScreen tests: ⚠️ Need Hive/Provider mocking
- Repository: ✅ Clean and ready for development

Ready for next agent to continue with provider testing and achieve 90% coverage goal.

+++ lib/widgets/recipe_card_stack.dart
@@ -137,8 +137,8 @@ class _RecipeCardStackState extends State<RecipeCardStack> {
                 },
                 child: RecipeCard(
                   recipe: recipe,
-                  horizontalThreshold: horizontalThreshold,
-                  verticalThreshold: verticalThreshold,
+                  horizontalThreshold: horizontalThreshold.toDouble(),
+                  verticalThreshold: verticalThreshold.toDouble(),
                 ),
               );
             },
